models/model/requirements.txt
numpy==1.20.3
pillow==8.2.0
tensorflow==2.6
